librbd: simplify IO method signatures for 32bit environments

+++ src/librbd/AioImageRequestWQ.h
@@ -20,13 +20,13 @@ public:
   AioImageRequestWQ(ImageCtx *image_ctx, const string &name, time_t ti,
                     ThreadPool *tp);
 
-  ssize_t read(uint64_t off, size_t len, char *buf, int op_flags);
-  ssize_t write(uint64_t off, size_t len, const char *buf, int op_flags);
+  ssize_t read(uint64_t off, uint64_t len, char *buf, int op_flags);
+  ssize_t write(uint64_t off, uint64_t len, const char *buf, int op_flags);
   int discard(uint64_t off, uint64_t len);
 
-  void aio_read(AioCompletion *c, uint64_t off, size_t len, char *buf,
+  void aio_read(AioCompletion *c, uint64_t off, uint64_t len, char *buf,
                 bufferlist *pbl, int op_flags);
-  void aio_write(AioCompletion *c, uint64_t off, size_t len, const char *buf,
+  void aio_write(AioCompletion *c, uint64_t off, uint64_t len, const char *buf,
                  int op_flags);
   void aio_discard(AioCompletion *c, uint64_t off, uint64_t len);
   void aio_flush(AioCompletion *c);
